fix the requested issues
