+++ src/components/NavBar.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import './NavBar.css';
 import { Link } from 'react-router-dom';
-import { FaUser } from 'react-icons/fa';
 
 const NavBar = () => (
   <nav className="menuNav">
@@ -16,7 +15,7 @@ const NavBar = () => (
         <li><Link className="menui vertical" to="/Categories">Categories</Link></li>
       </ul>
     </div>
-    <div className="imgAccount"><FaUser fa-lg /></div>
+    {/* <div className="imgAccount"><FaUser fa-lg /></div> */}
   </nav>
 );
